+++ test/main/testHelper.py
@@ -4,49 +4,33 @@
 
 import unittest
 
-from src.main.LabelGenerator import LabelGenerator
+from src.main.Helper import *
 
 
-class TestLabelGenerator(unittest.TestCase):
+class TestHelper(unittest.TestCase):
 	"""
-	TestLabelGenerator: Test class for LabelGenerator
+	TestHelper: Test class for Helper file
 	"""
 
 
-	def test_buildImageNormal(self):
+	def test_printVersions(self):
 		"""
-		Test if buildImage works when called normally
+		Test if printVersions works without exception
 		:return: void
 		"""
-		lg = LabelGenerator()
-		lg.buildImage("give_away", "Text", outputfile=None)
+		printVersions()
 
-	def test_buildImageGiveAwayWithEmptyText(self):
+	def test_getVersion(self):
 		"""
-		Test if buildImage works normally for GiveAway with empty text
+		Test if getVersion return a non-empty string
 		:return: void
 		"""
-		lg = LabelGenerator()
-		lg.buildImage("give_away", "", outputfile=None)
+		s = getVersion()
+		self.assertEqual( str, type(s) )
+		self.assertTrue(len(s) > 0)
 
-	def test_buildImageWrongLabel(self):
-		"""
-		Test if buildImage throws an error for unknown label type
-		:return: void
-		"""
-		lg = LabelGenerator()
-		with self.assertRaises(Exception):
-			lg.buildImage("beef", "Text", outputfile=None)
 
-	def test_buildImageUndocumentedWithEmptyText(self):
-		"""
-		Test if buildImage throws an error for labels different to
-		GiveAway and an empty text.
-		:return: void
-		"""
-		lg = LabelGenerator()
-		with self.assertRaises(Exception):
-			lg.buildImage("undocumented", "", outputfile=None)
+
 
 
 if __name__ == '__main__':
