+++ test/test_All.py
@@ -1,0 +1,13 @@
+
+# Author:    Maximilian Noppel
+# Date:      April 2021
+
+import unittest
+
+from main import *
+
+class TestAll(unittest.TestCase):
+	pass
+
+if __name__ == '__main__':
+	unittest.main()
